requirements.txt
pandas==1.1.5
numpy==1.19.5
matplotlib==3.3.4
boto3==1.20.45
pysurfline
configparser==5.0.2
psycopg2-binary
pendulum==2.1.2